# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 家族用パスワード管理システム API
  description: |
    家族内で安全にID・パスワード情報を共有・管理するためのWebアプリケーションAPI
    
    ## 認証
    - クライアント証明書による認証
    - セッショントークンによる認可
    - マスターパスワードによる暗号化キー生成
    
    ## セキュリティ
    - すべてのパスワードデータはクライアントサイドで暗号化
    - Zero-Knowledge Architecture
    - レート制限・ブルートフォース対策
    
  version: 1.0.0
  contact:
    name: Family Password Manager
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://family-pass.example.com/api
    description: 本番環境
  - url: https://staging.family-pass.example.com/api
    description: ステージング環境
  - url: http://localhost:3000/api
    description: 開発環境
security:
  - BearerAuth: []
  - ClientCertAuth: []
paths:
...
  /auth/session:
    post:
      tags:
        - Authentication
      summary: セッション作成
      description: クライアント証明書を検証し、新しいセッションを作成
      security:
        - ClientCertAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_info:
                  $ref: '#/components/schemas/DeviceInfo'
      responses:
        '200':
          description: セッション作成成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SessionData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    
    get:
      tags:
        - Authentication
      summary: セッション検証
      description: 現在のセッションの有効性を確認
      security:
        - BearerAuth: []
        - ClientCertAuth: []
      responses:
        '200':
          description: セッション有効
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    delete:
      tags:
        - Authentication
      summary: ログアウト
      description: 現在のセッションを無効化
      security:
        - BearerAuth: []
      responses:
        '204':
          description: ログアウト成功
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/master-password:
    post:
      tags:
        - Authentication
      summary: マスターパスワード検証
      description: マスターパスワードを検証し、暗号化キー生成用のソルトを返す
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - master_password_hash
              properties:
                master_password_hash:
                  type: string
                  description: クライアントサイドでハッシュ化されたマスターパスワード
                  example: "pbkdf2_sha256$100000$salt$hash"
      responses:
        '200':
          description: マスターパスワード検証成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MasterPasswordVerification'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # パスワード管理
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: セッショントークンによる認証
    ClientCertAuth:
      type: apiKey
      in: header
      name: X-Client-Cert
      description: クライアント証明書による認証
...
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - timestamp
        - request_id
      properties:
        success:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time
          example: "2024-06-23T12:00:00Z"
        request_id:
          type: string
          example: "req_abc123"
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
        - request_id
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time
          example: "2024-06-23T12:00:00Z"
        request_id:
          type: string
          example: "req_abc123"
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "AUTH_CERT_INVALID"
        message:
          type: string
          example: "クライアント証明書が無効です"
        details:
          type: object
          additionalProperties: true
        field:
          type: string
          description: バリデーションエラー時のフィールド名
...
    DeviceInfo:
      type: object
      properties:
        user_agent:
          type: string
          example: "Mozilla/5.0..."
        platform:
          type: string
          example: "iPhone"
        app_version:
          type: string
          example: "1.0.0"
    SessionData:
      type: object
      properties:
        session_token:
          type: string
          example: "sess_abc123..."
        expires_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        certificate:
          $ref: '#/components/schemas/Certificate'
    UserSession:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        session:
          type: object
          properties:
            created_at:
              type: string
              format: date-time
            last_accessed_at:
              type: string
              format: date-time
            expires_at:
              type: string
              format: date-time
        permissions:
          type: array
          items:
            type: string
          example: ["password:read", "password:write", "history:read"]
    MasterPasswordVerification:
      type: object
      properties:
        verified:
          type: boolean
          example: true
        encryption_salt:
          type: string
          example: "random_salt_for_encryption"
        key_derivation_params:
          type: object
          properties:
            algorithm:
              type: string
              example: "PBKDF2"
            hash:
              type: string
              example: "SHA-256"
            iterations:
              type: integer
              example: 100000
            salt_length:
              type: integer
              example: 32
    User:
      type: object
      properties:
        id:
          type: string
          example: "user-123"
        name:
          type: string
          example: "father"
        display_name:
          type: string
          example: "お父さん"
        role:
          type: string
          enum: [father, mother, son, daughter, admin]
        preferences:
          type: object
          additionalProperties: true
    Certificate:
      type: object
      properties:
        fingerprint:
          type: string
          example: "aa:bb:cc:dd..."
        expires_at:
          type: string
          format: date-time
        serial_number:
          type: string
          example: "1234567890"

    # パスワード関連
...
  responses:
...
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "AUTH_CERT_INVALID"
              message: "クライアント証明書が無効です"
...
    RateLimited:
      description: レート制限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "AUTH_RATE_LIMIT_EXCEEDED"
              message: "認証試行回数の上限に達しました"
              details:
                retry_after: 900
...